--- a/src/ballet/ed25519/Local.mk
+++ b/src/ballet/ed25519/Local.mk
@@ -1,4 +1,6 @@
 $(call add-hdrs,fd_ed25519.h)
 $(call add-objs,fd_ed25519_fe fd_ed25519_ge fd_ed25519_user,fd_ballet)
 $(call make-unit-test,test_ed25519,test_ed25519,fd_ballet fd_util)
+$(call make-unit-test,test_ed25519_signature_malleability,test_ed25519_signature_malleability,fd_ballet fd_util)
 $(call run-unit-test,test_ed25519,)
+$(call run-unit-test,test_ed25519_signature_malleability,)