… src/ballet/ed25519/test_ed25519_signature_malleability_should_fail.bin
… src/ballet/ed25519/test_ed25519_signature_malleability_should_pass.bin